SMP Support for raspberry pi 4b

+++ bsps/aarch64/raspberrypi/include/bsp/raspberrypi.h
@@ -42,7 +42,7 @@
  * @{
  */
 
-#define BCM2711_REG(x)           (*(volatile uint64_t *)(x))
+#define BCM2711_REG(x)           (*(volatile uint32_t *)(x))
 #define BCM2711_BIT(n)           (1 << (n))
 
 /** @} */
